+++ rust/reim_wrapper_dt_wasm/src/lib.rs
@@ -4,5 +4,7 @@ pub mod imports;
 pub mod malloc;
 pub mod abort;
 pub mod host_module;
+pub mod wasm_receive;
 pub use receive::*;
+pub use wasm_receive::*;
 pub use host_module::*;
